Add optional metadictionary in generate functions
If the "GenKit" key is absent in the dictionary, the generate
functions will add a dictionary with that key for the template to
use that include metainformation like `datetime` which contains
the time the tool was run and `templatePath` (if available)
that describes the file used for the template content. These can
be accessed from a template as `{{ GenKit.datetime }}` which may
be useful in comments in generated code.

Sending a dictionary as metaDictionary in the generate functions
will insert the contents to the autogenerated dictionary GenKit
adds.